Change edge persistence to use strings instead of resource ids.
The problem is that edges as specified in the catalog can contain
resources in them that aren't actually listed in the resources
list for that catalog. Mostly things like Class[foo] and the like.

Until I figure out what to do about that, I'm just going to shove
in the raw edge sources and targets into a table, link them with
a host, and be done with it.