Compute dependency edges, make catalog keys keywords
This changeset adds the ability to take a wire-format catalog, scan its list of
resources, and build a dependency graph. This includes a subtle semantic change
around the :edges attribute of a cmdb catalog. Previously, that attribute just
contained a list of containment edges. Now, that attribute contains a list of
edges where each edge has a :source, :target, and :relationship attribute. The
comments in the code go into much more detail, but in a nutshell the edge list
now contains several different types of edges other than containment.

While developing this, it was a good opportunity to convert the
wire-representation's use of strings for keys in the catalog's many maps, and
instead convert those to keywords. This makes for more concise lookup code, and
is generally more idiomatic.

As this is the first moderately complicated code in the project, I've added
documentation. Docs for the parsing routines have been written in Markdown
format, so as to be useful when using the Marginalia documentation generation
tool. You can generate the docs by using 'lein marg'.