Add back in basic deduping to allow main execution path to work
In earlier code, I removed the functions that computed a resource deduping hash
from the catalog parsing/transformation code. My rationale was that as that
code doesn't really have any "domain" value, and exists purely as a storage
optimization, it really belongs with the rest of the persistence code.

But because I'm dumb, I forgot that taking that code out would break the
current (crappy/prototype) main execution path. I've put some basic dedupe code
back in, though this is purely a stop-gap to get things working again. Better
dedupe is coming.