Add support for storing SCF data in H2 database
I've migrated the schema creation code into the scf storage layer itself, which
allows us to abstract that creation across databases we'd like to support (H2
and PG). I've added multimethods for the types of things that are
database-specific, which at this point is anything relating to SQL ARRAY types.

The tests for the storage code are all done using in-memory H2 databases, for
speed and easy testability without needing a bunch of external systems
configured.