Fix database detection in our multi-methods.
When we moved from the raw connection to the connection pool, we lost the
ability to dispatch to the right database-specific implementation based on the
direct class of the connection object.

Instead, use the database product name from the metadata.  This is somewhat
less instantly direct, but is correct regardless of how many layers we end up
wrapping the underlying connection in.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>