Migrate H2 => HSQLDB for array column query support.
Turns out that H2 has array columns, but they are completely useless for any
sort of query operation.  Specifically, you can only have complete equality of
the array as a query operation.

HSQLDB, while not quite PGSQL, is able to perform a "member of" query
against an array, which means it meets our standards for compatibility testing
and performance.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>