Externalize DB/MQ configuration to INI file
This changeset changes the CMDB, causing it to load in parameters for resources
like databases, message queues, and web servers to an external file (so we can
finally get rid of all the hard-coded stuff scattered around our source files).
This was a little more involved than I originally thought; to get this working,
I had to do the following:

1) The with-scf-connection code relied on global state to work, and that global
state was hard-coded at the top of storage.clj. I moved the
database-pool-related code into com.puppetlabs.jdbc, where I also added a
utility function that produces a clojure.java.jdbc-compatible, pool-backed
datasource. This is essentially what with-scf-connection was doing for us, but
I put that functionality in a more neutral spot. And because we're now just
dealing with vanilla java datasource objects, we can use the vanilla
with-connection func to provide database access when we need it. The resource
querying code was already very well-encapsulated, so there were only one or two
spots where I had to make changes to make all of that work again.

2) In order to configure things from an external configuration file, we need to
have a main entry-point into the app that can take the path to the
configuration file as an argument. This is easy enough to do for the
command-handling part of the CMDB, but the ring app is trickier; we're using
lein-ring, and its ring-init func is forbidden from taking any arguments. When
we move to using a proper servlet container, we can piggy-back on what it's
doing for configuration. But since we're not there yet, we need another entry
point other than lein-ring that we can hook into. As a stop-gap, I've created a
vanilla CLI-based way to execute the webapp.

3) I've created a ring-middleware function that wraps every request with
"global" state. This is done to avoid having top-level vars at the namespace
level that contained configuration information. Now, we can configure the
webapp up-front, and any request handler can just look in the request map to
figure out how to get access to whatever external resources it needs.
Currently, the only such resources we're adding to the request is the
connection information for the SCF database.

4) Because the ring app now needs to be configured, some of the test-cases that
assumed that something else would configure the ring app had to be refactored
to use a single, pre-configured application instance per-test.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>