Switch custom exceptions from records to vanilla maps
'defrecord' is used to create a proper Java class, and is primarily useful for
Java interop. This creates some friction when trying to use a record in another
namespace, as you have to treat the record the same way you'd treat Java
classes used for interop. Since the exception handling library we're using can
throw/catch arbitrary objects, there's little point in using a heavyweight
record when a standard map would do.

This has a few benefits:

1) It's more idiomatic clojure
2) It makes for prettier code
3) It doesn't run afoul of compilation/classloader problems

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>