Upgrade our version of org.clojure/tools.cli
They changed the API slightly since the (admittedly ancient) version we were
previously using, so some slight refactoring was in order. The new versions of
the library defer printing help to the caller, so I went ahead and wrote a
utility function that takes care of adding help switches, printing help text,
and ending the process when necessary.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>