Add an API for querying facts
This adds the route /facts/:node, which returns a JSON hash of the facts
for a node. The returned hash contains two keys, "name" and "facts".
"name" is the name of the node, and "facts" is a hash from fact name to
fact value. Because this route logically maps a "facts" resource, it
will respond with 404 when a certname is present, but no facts are known
for it. Naturally, it will also respond with 404 when the certname isn't
present.