Add support for fact persistence
This adds a handler for the "replace facts" command. It expects a
payload which is a map with keys for "name" and "value". "name" is the
name of the node and "value" is itself a map from fact name to fact
value.

Facts are stored in the certname_facts table as a simply key/value pair,
with an additional column for the associated certname.