Move from data.json to cheshire for JSON parsing
Cheshire is much, much faster. The port was pretty mechanical, barring a few
gotchas:

1) The method names are slightly different

2) By default, data.json will automatically turn JSON hash keys into keywords,
unless you specify an additional boolean argument of "false". Cheshire defaults
to parsing the JSON as-is, and only keyword-ifies keys when you supply a
boolean argument of "true". So porting required some argument swizzling.

3) data.json would throw an EOFException when passed in an empty string, but
Cheshire simply returns nil. A test case had to change to accomodate this
situation. In the course of modifying that test case, I found that the
post-conditions for one of our catalog-parsing methods was wrong, and had to
fix it to get the method to properly throw an AssertionError.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>