Use an embedded ActiveMQ for command-processing
This patchset embeds a fully-persistent ActiveMQ broker within Grayskull. For
embedded purposes, we're using the "vm://" transport. This makes embedding
easy, and prevents malicious actors from connecting to the (currently
internal-use-only) MQ.

As only services within the same process can communicate with the MQ, this
patchset also creates a single entrypoint for grayskull that sets up all the
main components for execution within the same VM.

- A new CLI entrypoint, services.clj, that sets up the MQ, command processor,
  and REST server

- The command-line entrypoints contained in commandproc.clj and queryserver.clj
  have been subsumed by services.clj, and are thus obsolete.

- A new MQ namespace that provides a few utility functions for interacting with
  embedded MQs.

- Changes to the sample configuration file, removing MQ-related variables that
  we no longer need (mq address, mq endpoint) and adds a configuration variable
  for the directory to use for MQ persistence.

- A new test helper macro, "with-test-broker", that handles setup/teardown of
  an embedded MQ. This makes writing tests that interact with an MQ much
  easier.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>