Store parameter values as serialized objects, not arrays
Previously, we used an array column to store parameter values. This
required some db abstraction code to handle it in the various
db-specific ways, while not really providing significant value (and not
handling hash parameters).

Now we just serialize parameter values to store them in a text
field. Similarly, to query against the serialized field, we serialize
the value we're looking for.

The serialization is an ordinary JSON dump of the object, except for
maps, which are first maps so equivalent objects serialize the same.