Use ClojureQL resource '=' queries, and retrieving resource/params
The primary material benefit of ClojureQL is that it can handle "in"
operations, which is necessary when we need to select resources and
parameters from a set of hashes.

Additionally, it lets us write less of our own unsightly SQL code.
Unfortunately, it can't handle the complex joins we need for and/or/not
queries, so we still have to construct those manually, but it provides
useful primitive building blocks with which we can develop higher order
operations.