Add whole-catalog deduplication
If we receive a catalog for a host that contains the same resources,
parameters, edges, classes, tags, etc, and only the catalog's metadata has
changed, then we don't actually have to store the new catalog; we can simply
update the metadata. If the common case in the field is that catalogs are
compiled much more often than they actually change, then this can result in a
substantial performance improvement.

* Create a separate table for just certnames, distinct from catalog metadata.
  This allows for the catalog schema to evolve independent of the notion of a
  "host" in Grayskull. It also allows other, non-catalog data to share the list
  of hosts.

* Add operations for manipulating catalogs and certnames independently.

* Add an algorithm for computing a "similarity hash" for a catalog, that is, a
  hash that uniquely identifies a catalog for the purpose of storage (does it
  have the same resources, tags, classes, etc).

* Change the replace-catalog command to add a certname if necessary.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>