Fix clojureql's shadowing of clojure.core functions
This causes a problem with a persistent compilation process, like swank,
wherein it can't determine which is the correct implementation of a function
that is imported from 2 different namespaces without a prefix. Since clojureql
uses function names like "conj!" and "distinct" that also exist in
clojure.core, there's no good way for the compiler to determine which function
is being referred to in the code.

I've implemented a temporary fix, which is to simply exclude all overlapping
functions from clojure.core.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>