Index the certname_facts table appropriately
This indexes certname_facts on both `certname` (for fetching facts for a given
node) and `fact` (for fetching the values of a given fact).  Also changed the
primary key of certname_facts to be (certname,fact), as a given certname should
only have a single value for a given fact.