Implement basic schema migrations
The database can be migrated using the `migrate!` function, which will
determine the necessary migrations to apply and run them. A migration is
added by defining a function in migrate.clj, and adding it to the
`migrations` map. The database will be automatically migrated whenever the
`services` or `populatemq` CLI applications are run.

Migrations are recorded in the schema_migrations table, with a row
containing the version number and the time at which the migration was
run.

The `migrations` and `pending-migrations` functions, along with the
schema_migrations table, may be useful to provide introspection about
the current state of the database.