Command REST endpoint should be more lenient with content-types
Clothesline is very literal when you specify what content-types you accept. If
the strings don't match exactly, it rejects the incoming request with a 415
error. If we happened to get a content-type that looked like:

    application/x-www-form-urlencoded ; charset=UTF-8

...then it would get rejected, as we hadn't considered the trailing charset
definition. We relax this restriction, and now allow for content-type strings
that include charset information.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>