Use statement batching when inserting resources
We would previously do several SQL statements for insertion of a resource:

* insert its metadata
* N inserts for all of its parameters
* N inserts for all of its tags
* an insert to associate it with a catalog

...and we'd do that for all N resources in a catalog. That's a lot of queries.
Now, we use statement batching to queue up all the pending inserts on a
per-table basis. We then execute them in foreign-key-satisfying order:

* insert metadata for all resources in the catalog
* insert all resource parameters for all resources
* insert all resources tags for all resources
* associated all of the resources with the catalog

We do this all in a single round-trip to the database, saving a ton of wasted
time doing back-and-forth communication.

Note: this patch does not remove the query-per-resource penalty we pay when
determining if a resource exists. Even though that penalty remains, batched
inserts are still a pretty massive performance boost.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>