Explicitly load database driver class when building the connection pool
When executing via leiningen, the order of class loading is slightly different
than when executing via JAR. The canonical way of initializing a JDBC driver is
to call (Class/forName ...) on it, therefore making the JDBC driver management
machinery aware of it. We weren't explicitly doing this, and were lucky that
execution via leiningen worked.

This patch makes the loading of the driver class explicit, thereby Doing The
Right Thing.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>