Restructure the way resources are stored
We now deduplicate only the attributes of a resource involved in
managing that resource. So now multiple catalogs that define equivalent
resources will share the same resource, regardless of metadata including
file, line, tags, or exported-ness. A resource's hash is now computed
with only type, title, and its parameters.

Metadata now lives on the catalog_resources table, because it's specific
to the particular instance of the resource in a given catalog. This
includes tags (which are now stored as an array), so the resource_tags
table is gone. Type and title are also moved to the catalog_resources
table, so the resources table is gone as well.

The removal of the resources table means there's no longer a foreign key
relationship for edges or resource_params, so those need to be
garbage collected.