Replace Clothesline with Moustache
Clothesline is orphaned code, and we don't currently have the bandwidth to
maintain it. Given that our current HTTP needs are extremely basic (we do very
little content negotiation, we don't have complicated REST entities, we aren't
generating etags, etc), we can go without a proper REST-ful state machine for
now. It's not great, but I figure we can cross that bridge when we come to it
(all of our HTTP handling logic put together is < 200 LOC).

* Changed the REST endpoint state machines to simple Ring apps

* Changed our exotic content-types for facts and queries to something mortals
  can type

* Changed the error responses in several cases from JSON to plain text. This is
  because, according to Webmachine, malformed URL checks happen prior to
  content negotiation, and hence are okay to send back in plain text.

* Changed one of the macros in the resources test suite to a method; it didn't
  need to be a macro, and that just made debugging more complicated.

* Wrote a small utility function that does very basic content-negotiation

* Wrote a small utility function that helps generate JSON Ring responses

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>