Eliminate dependency on Midje
We were only using a single function from midje, and we didn't really need to
be. I rewrote the affected code and removed the dependency.

One side-effect of this was that the compiler turned up an error where we were
using a library without doing a 'require on it. It worked before because midje
included that library; without midje, we need to add an explicit 'require
(which we should be doing anyways).

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>