Move catalog randomization functions out of test/ into src/
When packaging Grayskull into a JAR, the test/ directory is intentionally
excluded by the build system. However, our benchmark code relies on some
functions from the test namespace and hence won't work when run from a JAR.

This patch takes the functions we were using and moves them into their own
namespace inside of src/. This was listed as a TODO on those functions anyways,
and this is an opportune time to do it.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>