Ring middleware that computes metrics for all HTTP requests
(wrap-with-metrics ...) will tack performance counters for each URI encountered
by the wrapped Ring app. Users supply a function that tells the middleware how
to "normalize" a URI; in doing so, the mapping of what URIs correspond to which
metrics is entirely customizable.

For each normalized path, we track the time required to service a request as
well as independent meters for each status code.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>