Move from sync to async MQ consumers
The new version of clamq brings some good absractions for asynchronous MQ
consumers. These has some desirable properties when compared to the existing
synchronous interface:

The sync interface creates a lazy-sequence on top of the MQ, which can be
tricky. Lazy sequences are "chunked" by Clojure under-the-hood as an
optimization, buffering up 32 items at a time. Normally that's great, but when
each item in the lazy sequence is a several-MB-large message, that uses a fair
amount of memory. This is something that ActiveMQ handles on its own anyway, in
that it does its own pre-fetching inside of Consumer code. I don't think the
complexity buys us much of anything.

The other problem is that with sync consumers, we had to do explicit ack-ing of
messages in our code.

Async consumers have a much simpler lifecycle: a message comes in, a callback
in invoked. If the callback succeeds, the message is acked. That's pretty much
it. It allows us to better structure the code in terms of concentric handler
functions with predicatble lifecycles.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>