Much faster string hashing
We previously didn't bother specifying an encoding to use when converting a
string to bytes. This causes the JVM to use the platform's default encoding,
which has several drawbacks:

* The default platform encoding could be really slow (like, say, MacRoman)

* Switching a Grayskull instance from one platform to another would result in
  completely different hashes for the same input

This patchset specifies UTF-8 as the target encoding when computing a hash.
This makes Grayskull more portable, and the hash computation on my laptop is
60% faster than before.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>