Lookup the appropriate digest method once
Looking up the multimethod caused up to 10x slower performance for the
digest operation. Since we don't ever call it with different types of
arguments anyway, we can exploit that for a significant performance gain
(given that we call this function once per resource) by fetching the
appropriate method once, and then using that.