Command-processing middleware for better thread names
Spring's MQ listener classes (used internally by clamq) stupidly hard-code
their thread name to the name of the class itself. Normally this isn't a big
deal, but if you want to have multiple, simultaneous listeners they all end up
having the same name ("DefaultMessageListenerContainer-1" or something equally
ridiculous). It makes it appear in the logs that there's only a single thread,
when there may actually be several.

This middleware simply renames the current thread such that it has a meaningful
prefix. To ensure the thread names are unique, the JVM-internal thread ID is
appended to each thread's name.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>