Change command payload to a JSON-stringified object
Using a stringified object here lets us separate command parsing from
payload parsing. This means we can be more-or-less guaranteed that at
least the top-level parse of a message will succeed, regardless of its
payload. Parsing this separately lets us refer to the command and version of
the message when logging failures that occur when parsing/handling the
command payload itself.