Add an :annotations element to commands
This element isn't required to be specified when submitting a command, but if
it is, it is a map containing arbitrary keys and (JSON-serializable) values.
This can be used to store information that arises during the lifetime of a
command.

Currently we only use annotations to store the "retries" for the command. This
actually represents an attempt at processing the command, and is stored as a
list of the timestamps at which the command was parsed.