Implement a "dead letter office" for failed messages
Failed messages are now annotated with an entry in their :attempts array
before being republished to the queue. The attempt is recorded as a map
with keys :timestamp, :error, and :trace, which are the time at which
the failure occured, the error message, and the stack trace (as an array
of strings) from the exception.

When a fatal error occurs (including parse errors) or when the message
has exceeded its allowed retries, it will be discarded to the dead
letter office.  Each failed message is stored in its own file at
<mq-data-dir>/discarded/<command>/<timestamp>-<checksum>. This file
contains the message itself in JSON form (in its final state, including
all "attempt" annotations) plus the long-form exceptions/traces from
each attempt.