Add resource namevars as aliases if not accounted for
Resources can require/notify/etc other resources based on their name,
when the referenced resource has a different title, causing an edge not
to be properly created. Now we add the namevar as an alias so these
resources can be found and edges created appropriately.

Paired-With: Deepak Giridharagopal <deepak@puppetlabs.com>