Implement conversion to UTF-8 when submitting data to Grayskull
On Ruby 1.8, this will assume all input is UTF-8-ish, and will drop
invalid UTF-8 byte sequences. On Ruby 1.9, it will properly convert, as
long as the original encoding is correctly known, otherwise it will
replace invalid or undefined characters with the unicode replacement
character \ufffd. In all cases, if bytes are dropped, a warning will be
issued.