Fix namevar alias creation for resources
The alias creation code sets alias using the string 'alias', rather than
the symbol :alias, but the code that uses the aliases to lookup
resources for edge creation was using :alias. This caused the old set of
aliases (that was provided by to_pson_data_hash) to be used incorrectly.