Support "swank" protocol for remote REPL
Currently, we only support the "nrepl" protocol for remotely connecting to
Grayskull for live debugging. This is nice, but it's meant for connections via
the command line, using telnet. Every clojure IDE has the ability to connect to
a remote program and manipulate it in realtime, so why can't Grayskull support
that?

This patchset adds a "type" flag to the set of remote REPL options. We default
to "nrepl", but you can also specify "swank". Swank is the protocol Emacs uses
to talk to a remote lisp, and it supports all kinds of awesomeness like live
code-completion, cross-referencing, sending single functions (or an entire
namespace) to a running clojure instance thereby redfeining methods on-the-fly,
etc.

You can connect to a running Grayskull instance from within Emacs by doing:

    M-x slime-connect

This works, is a little mind-bending, and is completely awesome.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>