Use exponential backoff when retrying commands
Previously, we'd immediately re-inject a command back into the MQ if there was
a non-fatal error when processing that command. If the reason for the error was
some kind of transient problem like, say, a database was down or a disk is
full, then it's possible we'd end up retrying a command N times, back-to-back,
in less time than it would take to troubleshoot the transient problem.

Rather than retry immediately, this patchset implements an exponential backoff
algorithm: 8^N seconds, where N is the current attempt number. We'll still
permanently fail a message if it exceeds the limit of allowed attempts.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>