Add ability to filter resources by node active-ness
This adds a query term `["node" "active"]` which accepts true or false,
returning resources only for nodes which are active or inactive,
respectively. If omitted, resources for both active and inactive nodes
will be returned. The resource terminus for storeconfigs searches
filters its queries to only active nodes.

This also changes the resource query term for node names to `["node"
"name"]` instead of just `"node"`.