Add `received` annotation to messages when parsing if necessary
The command processor is expecting messages to have been annotated with
a `received` timestamp by the HTTP endpoint enqueuing them. However,
this isn't necessarily going to be true for messages that get into the
queue other ways (such as tests), so now the processor will add the
timestamp if it's missing when parsing the command.