Quiet errors from initial command retry attempts
There are cases where a command may cause an error, yet the error is harmless
and the command will succeed upon retry a few seconds later. We've received
feedback from users that the messages displayed for transient errors are more
scary than helpful.

This patchset uses the current attempt # to determine the level at which errors
are logged during retry. For the first several retries, the error causing the
retry is logged to DEBUG. If, after several retries the command is still
causing problems, we log the error to ERROR. This way the errors users most
often see are likely indicative of real problems.

It's worth noting that no information is lost in this patchset; every command
is annotated with the full exception for *every* retry, that information is
preserved and represented in the Dead Letter Office of failed commands, and if
the user truly wishes to see those errors for every attempt it's a one line
change to their logging configuration (and that change can be applied live,
without taking Grayskull down).

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>