Use bulk query to determine resource existence
Previously, we used a SELECT for each resource we wanted to store to determine
if it already exists in the database. If a catalog has 1000 resources, this
results in 1000 queries. It's much more efficient to use a single query to
determine the existence of a set of resources in one shot.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>