Store receipt timestamp for catalogs
This allows us to make smart decisions about storing catalogs with
respect to retries. If the catalog in the database is more recent than
the catalog we are being asked to store (which is most likely to happen
if the current command has been retried, and another catalog was stored
in the meantime), we will simply ignore the command.