Add with-transacted-connection macro, use that everywhere
With postgresql, you can never truly operate outside of a transaction (which is
a good thing); You get one automatically when you open a new connection. If we
issue a malformed query, and if the query isn't inside of a (sql/transaction)
block, we never explicitly call "rollback" upon an exception (which you need to
do to "clear" the aborted transaction). From that point on, that database
connection is in a bad state; the database will reject any subsequent commands
until a "rollback" is issued (which won't happen).

Theoretically, the connection pool will eventually discover that the connection
is poisoned and will kill it off. While that's a nice safeguard, it doesn't
happen immediately and in the meantime, all queries over a poisoned connection
will fail. We can proactively prevent this by properly enclosing all database
activities within explicit transaction blocks.

This patchset adds a new macro, `with-transacted-connection`. It's exactly like
`with-connection`, except that it automatically executes enclosed forms within
an explicit transaction.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>