Use clj-time functions rather than standard Java objects
We were previously creating a java.util.Date instance to get the
current time. Because we're trying to avoid using standard Java date
libraries in favor of Joda Time (via clj-time), it's better to use the
function that clj-time provides us. Also, the function `now` is more
semantic than the implicit behavior of `(java.util.Date.)` to use the current
time.