Add a column (and table) for facts timestamps
This introduces the new table certname_facts_metadata, which now exists
to join certnames and certname_facts, representing a "set" of facts.
Currently the only metadata is timestamp. Similarly to catalogs, a
"replace facts" command will be ignored if the current facts in the
database are newer than the submitted facts.