Update metrics-clojure so we can catch bad payloads
Previously, wrapping our calls to process-command! in (time!) was
removing the {:fatal true} context from them, causing them to be treated
as non-fatal errors, and thus retried rather than properly discarded.
Upgrading metrics-clojure fixes this.

Using the new version of metrics-clojure and metrics revealed that our
(report-to-jmx) function was unnecessary (and actually harmful), as the
metrics library already reports to JMX.