Externalize PuppetDB configuration to a file
This will read config from $confdir/puppetdb.conf, in ini format:

    [main]
    server = main_server
    port = 1234

If no value is specified for either option, the global defaults of
'puppetdb' and '8080' will be used.

Also changed the terminus constants from PuppetDB to Puppetdb, to behave
properly with Puppet's conversion between names and constants.
Unfortunately, the "name" of PuppetDB is puppet_d_b, which isn't right.