Add config handling to cli! function
The --config argument is used by all current applications, and will most
likely be used for any future ones as well. Therefore it makes sense to
extract handling the argument to the shared cli! function.

Also added top-level error handling, which will print any exceptions
which reach it. If --trace is specified, it will also print the
stacktrace.