Expose errors that occur during startup of components
Because these execute in futures, their exceptions were being silently
swallowed, and we would diligently continue on, unable to do anything.
Now they are each given a promise that they can deliver an error on,
which will be handled by the services application itself, cancelling
each future and shutting down the broker, then throwing it to the
top-level handler to be printed.