Extract configuration and command submission to a Util module
Given that we assume a puppet master is also running an agent, we can
rely on whatever util code we need being pluginsynced and loadable.
Thus, we don't actually need to use a hack with features to load our
config. Similarly, we can use methods to abstract the behavior related
to submitting commands.

This also introduces better error handling in the new submit_command
method, to address cases where the server is down or returns an
unsuccessful response.