Change command submission result to be a map
JSON only allows objects and arrays at top-level, so our JSON string is
not valid, which causes problems for consumers trying to parse it to get
the UUID. The response is now `{:uuid <uuid>}`. This form also allows us
to easily add more information in future, such as state of the queue,
etc.