Rebase init script for the one actually used
Previously I had grabbed the wrong init script for puppetdb.  It was an
older copy.  This one removes usage of daemonize.rb and complies with
debian standards.

This is largely taken from Ben with a few updates/changes to make it a
bit more generic and named properly for puppetdb.  I also updated the
default file which is used in /etc/default and /etc/sysconfig.  Since no
Red Hat init script is written yet, I overwrote the one with this new
debian one as a starting point.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>