Ensure there's at least 1 command processing thread
The default setting for the number of command processing threads is set to half
the number of cores in your system (rounded down due to integer division). That
works great until you've got a system with only one core, in which case we ended
up starting 0 command-processing threads. This patch ensures a minumim of one
command processing thread.

It's 2012...who's still rocking a single core? :)

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>