Fix tests for command-processing configuration to work on single-core systems
While a prior patchset fixes the main code for auto-configuring threads on a
single-core system, it didn't fix the test cases. So if you ran the tests on a
single core system, they'd fail.

This patchset abstracts out the code that determines the number of cores in a
host into a utility func, and test cases override that func during verification
of behavior.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>