Disable default resource queries
If you don't supply a query parameter to the /resources/ REST endpoint, we
return *all* resoures with *all* parameters for *all* hosts. That, as you may
guess, is a horribly expensive operation.

This patch disables the default query, and mandates that a query is explicitly
supplied by the caller. Plus, that's actually the behavior that we documented
in our specs...so we now better deliver on our promises. :)

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>