Smarter query compiler
Rather than compiling each "=" operation independently and joining them
with UNION/INTERSECT/JOIN, we now build a list of required tables, and
WHERE clauses. "and"/"or" operations simply combine their child WHERE
clauses with AND/OR respectively. "not" will still evaluate its children
as an "or" operation, and will then wrap that clause in NOT.

This removes all unnecessary set operations and joins, and also
unnecessary subqueries, resulting in tremendous performance gains across
the board.

The only subqueries we produce now are EXISTS queries for
resource_params queries, to avoid joining n times to resource_params in
the case of "and" operations.