Don't use a subselect to get resources
We were previously constructing a query which would return a set of
catalogs and resources, and were then querying for the columns we
cared about from catalog_resources matching in that set. However, now
we're just constructing a set of JOINs and a WHERE clause, so we can
append that to the query that retrieves the columns we want.

Also removed the ability to omit the query and retrieve all resources,
since that capability has been removed from the HTTP endpoint.