Delete resource query tests asserting SQL output
These have become laborious to maintain, as any changes to query
generation cause basically all of them to need to be changed. Given that
they will start failing on effectively *any* change to the code (correct
or not), they aren't providing any value in identifying regressions.
The end-to-end tests against a database which assert results are much
more useful.