Remove unnecessary join condition for certname_catalogs
Now that we don't use a subselect when querying resources, we're
*always* going to join against certname_catalogs in our main query, as
it has columns we always want to select. So it's not necessary to add
another join. Also, this code was broken anyway, as the
certname_catalogs condition was always returning an empty string for the
join expression.

The only legal table remaining to join on is certnames, which is still
optional, as it's only involved in the query if we're querying based on
node activity.