Fix documentation generation
The lein-marginalia plugin is stupid, and instead of snagging things like the
project name, description, and version from function arguments, it simply
defers to marginalia to parse the project.clj file. This is dumb, because
there's no need to parse that file at all...the whole point of running as a
lein plugin is that you get access to the project file, post-any-parsing.
Furthermore, the marginalia parser can't cope with the special code we have in
our project.clj file that auto-determines the version number by looking at git
tags.

This patchset simply inserts our own "lein marg" plugin that Does The Right
Thing with arguments prior to calling the stock plugin.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>