Add bounded-drain-to-vec! mq function
Test cases that relied on retrieving N messages from the MQ were previously
using the drain-into-vec! function, which took a timeout. If the timeout was
reached prior to receiving N messages, the function would simply return.

If the MQ is fast enough, this never actually occurs in practice when running
our test cases. If the MQ is slow, though (like on our CI systems), then the
timeouts could be reached much more often.

This patchset creates a new function, bounded-drain-to-vec!, that only returns
once the supplied number of messages has been drained from the MQ. This avoids
any time-based race conditions.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>